fix: Add NUMERIC type casts to ROUND functions for PostgreSQL compatibility
- Cast all ROUND function arguments to NUMERIC type
- Fixes 'function round(double precision, integer) does not exist' error
- Applied to all migration files (005, 006, 007)

+++ supabase/migrations/005_performance_indexes.sql
@@ -148,7 +148,7 @@ SELECT
     COUNT(DISTINCT r.id) FILTER (WHERE r.status = 'rejected') as rejected_requests,
     COUNT(DISTINCT r.doctor_id) as unique_doctors,
     COUNT(DISTINCT rv.id) as total_reviews,
-    ROUND(AVG(rv.rating), 2) as avg_rating,
+    ROUND(AVG(rv.rating)::NUMERIC, 2) as avg_rating,
     MAX(r.created_at) as last_request_at
 FROM pharmacies p
 LEFT JOIN requests r ON r.pharmacy_id = p.id
